+++ Ranger/UI/TourneyTable/TourneyTableView/Cells/Stack/StackCellView.swift
@@ -10,7 +10,7 @@ import Foundation
 import SwiftUI
 
 
-class StackCellView: PlayerViewModelCellView
+class StackCellView: PlayerCellView
 {
 
     
@@ -35,14 +35,14 @@ class StackCellView: PlayerViewModelCellView
         setNeedsDisplay(self.bounds)
     }
     
-    override func setup(with playerViewModel: Model.Player, in tableColumn: NSTableColumn?)
+    override func setup(with player: Model.Player, in tableColumn: NSTableColumn?)
     {
         // Checks.
         guard let column = tableColumn else { return }
         guard let textField = self.textField else { return }
         
         // Retain data.
-        self.textFieldData = playerViewModel.textFieldDataForColumnIdentifiers[column.identifier.rawValue]!
+        self.textFieldData = player.textFieldDataForColumnIdentifiers[column.identifier.rawValue]!
         
         // Apply text.
         textFieldData.apply(to: textField)